organise commands into modules
