+++ ModdedMinecraftClub.MemberBot/ModdedMinecraftClub.MemberBot.Bot/Modules/StaffCommandsModule.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
@@ -8,118 +6,8 @@ using ModdedMinecraftClub.MemberBot.Bot.Models;
 
 namespace ModdedMinecraftClub.MemberBot.Bot.Modules
 {
-    [SuppressMessage("ReSharper", "StringLiteralTypo")]
-    public class MemberApplicationsModule : ModuleBase<SocketCommandContext>
+    public class StaffCommandsModule : ModuleBase<SocketCommandContext>
     {
-        #region Regular
-
-        [Command("pending", RunMode = RunMode.Async)]
-        public async Task Pending()
-        {
-            using var c = new DatabaseConnection();
-            var list = c.GetAllPending();
-
-            if (!list.Any())
-            {
-                await Context.Channel.SendMessageAsync("There are no pending applications at the moment.");
-            }
-            else
-            {
-                var s = list.Aggregate("Pending applications:\n-------------------------------------------\n", 
-                    (current, app) => current + $"[{app.AppId}] {app.AuthorName}; {app.AppTime}\n\n");
-
-                await Context.Channel.SendMessageAsync($"```{s}```");
-            }
-        }
-        
-        [Command("approved", RunMode = RunMode.Async)]
-        public async Task Approved()
-        {
-            using var c = new DatabaseConnection();
-            var list = c.GetLast20Approved();
-
-            if (!list.Any())
-            {
-                await Context.Channel.SendMessageAsync("You haven't approved any applications yet.");
-            }
-            else
-            {
-                var s = list.Aggregate("Last 20 approved applications:\n-------------------------------------------\n", 
-                    (current, app) => current + $"[{app.AppId}] {app.AuthorName}; {app.AppTime}\n\n");
-
-                await Context.Channel.SendMessageAsync($"```{s}```");
-            }
-        }
-        
-        [Command("rejected", RunMode = RunMode.Async)]
-        public async Task Rejected()
-        {
-            using var c = new DatabaseConnection();
-            var list = c.GetLast20Rejected();
-
-            if (!list.Any())
-            {
-                await Context.Channel.SendMessageAsync("You haven't rejected any applications yet.");
-            }
-            else
-            {
-                var s = list.Aggregate("Last 20 rejected applications:\n-------------------------------------------\n", 
-                    (current, app) => current + $"[{app.AppId}] {app.AuthorName}; {app.AppTime}\n\n");
-
-                await Context.Channel.SendMessageAsync($"```{s}```");
-            }
-        }
-        
-        [Command("view", RunMode = RunMode.Async)]
-        public async Task View(int applicationId)
-        {
-            using var c = new DatabaseConnection();
-            
-            var app = c.GetById(applicationId);
-
-            if (app is null)
-            {
-                await Context.Channel.SendMessageAsync($":x: Application with ID `{applicationId}` does not exist.");
-                    
-                return;
-            }
-            
-            var b = new EmbedBuilder();
-            b.AddField($"{app.AppStatus.ToString().ToUpper()}: Application by {app.AuthorName}", $"Author's Discord ID: {app.AuthorDiscordId}\nApplication ID: {app.AppId}");
-            
-            if (app.MessageContent is null || app.MessageContent.Equals(""))
-            {
-                b.AddField("Provided details","*Player did not provide any details.*");
-            }
-            else
-            {
-                b.AddField("Provided details", app.MessageContent);
-            }
-
-            b.AddField("Link to original message", app.MessageUrl);
-            b.WithThumbnailUrl(app.ImageUrl);
-            b.WithFooter($"Applied at {app.AppTime}");
-                
-            switch (app.AppStatus)
-            {
-                case ApplicationStatus.Approved:
-                    b.WithColor(Color.Green);
-                    break;
-                case ApplicationStatus.Rejected:
-                    b.WithColor(Color.Red);
-                    break;
-                default:
-                    b.WithColor(Color.Blue);
-                    break;
-            }
-                
-            await Context.Channel.SendMessageAsync("", false, b.Build());
-        }
-
-        #endregion
-        
-        #region Staff Only
-
         [Command("approve", RunMode = RunMode.Async)]
         [Priority(1)]
         [RequireUserPermission(GuildPermission.BanMembers)]
@@ -213,7 +101,5 @@ namespace ModdedMinecraftClub.MemberBot.Bot.Modules
                 
             await Context.Channel.SendMessageAsync($":white_check_mark: **Rejected** application with ID `{applicationId}`");
         }
-
-        #endregion
     }
 }
